install

+++ jsp-demo-scaffolding-template-main/src/main/java/scaffold/framework/demo/repository/StudentRepository.java
@@ -1,8 +1,9 @@
 package scaffold.framework.demo.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaRepository ;
 
-import scaffold.framework.demo.entity.Student;
+import scaffold.framework.demo.entity.Student ;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
 }
+
